refactor: 테스트 코드 추가 및 패키징 계층화
