+++ src/main/java/wooteco/subway/line/exception/LineException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.line.exception;
+
+public class LineException extends RuntimeException {
+    public LineException(String message) {
+        super(message);
+    }
+}
