+++ src/main/java/wooteco/subway/line/exception/LineExistenceException.java
@@ -1,6 +1,6 @@
-package wooteco.subway.line;
+package wooteco.subway.line.exception;
 
-public class LineExistenceException extends RuntimeException {
+public class LineExistenceException extends LineException {
     private final static String message = "존재하는 노선 이름입니다.";
 
     public LineExistenceException() {
