+++ src/main/java/wooteco/subway/station/exception/StationExistenceException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.station.exception;
 
 public class StationExistenceException extends RuntimeException {
     private static final String message = "존재하는 역 이름입니다.";
