+++ src/test/java/wooteco/subway/line/LineServiceTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
+import wooteco.subway.line.exception.LineExistenceException;
+import wooteco.subway.line.exception.LineNotFoundException;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -46,6 +48,6 @@ class LineServiceTest {
         String sql = "insert into Line (name, color) values (?, ?)";
         jdbcTemplate.update(sql, "2호선", "초록색");
         assertThatThrownBy(() -> lineService.findById(2L))
-                .isInstanceOf(LineExistenceException.class);
+                .isInstanceOf(LineNotFoundException.class);
     }
 }
